changed versioning & fixed cleanup

+++ src/jansvbkernel/kernel.py
@@ -44,4 +44,5 @@ class jansvbkernel(Kernel):
                }
 
     def do_shutdown(self, restart):
-        shutil.rmtree(workingdir)
+        if os.path.exists(workingdir):
+            shutil.rmtree(workingdir)
